perf(web): convert /skills page to Server Component with ISR
- page.tsx becomes a Server Component that fetches /api/skills at
  revalidate=3600. Users see the full skill list in the initial HTML,
  no "Chargement des compétences..." spinner on first visit.
- Extracted the search/filter UI into SkillsClient. Language switch no
  longer triggers a re-fetch since skills payload already includes
  both nameFr/nameEn and description/descriptionEn.
- Ruleset toggle updates the URL so the server re-renders from the ISR
  cache instead of kicking off a client fetch.
